fix: UI交互优化
