+++ src/views/system/components/kindlemail.vue
@@ -1,25 +1,20 @@
 <template>
-  <div class="container">
-    <Breadcrumb :items="['menu.system', 'menu.system.smtp']" />
-    <div class="wrapper">
-      <a-spin dot :loading="dataLoading">
-        <a-row>
-          <a-col :span="20" :offset="2">
-            <a-form :model="form" size="large" :style="{ width: '700px' }" auto-label-width @submit="handleSubmit">
-              <a-form-item field="email" label="邮箱地址" required>
-                <a-input v-model="form.email" placeholder="请输入收邮件用的邮箱地址" />
-              </a-form-item>
-              <a-form-item>
-                <a-button html-type="submit" status="success" type="primary">{{
-                  $t('system.form.save')
-                  }}</a-button>
-              </a-form-item>
-            </a-form>
-          </a-col>
-        </a-row>
-      </a-spin>
-    </div>
-  </div>
+  <a-spin dot :loading="dataLoading">
+    <a-row>
+      <a-col :span="20" :offset="2">
+        <a-form :model="form" size="large" :style="{ width: '700px' }" auto-label-width @submit="handleSubmit">
+          <a-form-item field="email" label="邮箱地址" required>
+            <a-input v-model="form.email" placeholder="请输入收邮件用的邮箱地址" />
+          </a-form-item>
+          <a-form-item>
+            <a-button html-type="submit" status="success" type="primary">{{
+              $t('system.form.save')
+              }}</a-button>
+          </a-form-item>
+        </a-form>
+      </a-col>
+    </a-row>
+  </a-spin>
 </template>
 
 <script lang="ts">
