+++ src/views/system/index.vue
@@ -7,4 +7,6 @@
   </div>
 </template>
 
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+
+</script>
